Add files via upload
I have made a google forms asking "What is your name?" & "What is your e-mail?
I then filled this in 17 times using random names/emails. 
The code reads the link and downloads the responses in a temporary csv file.
After the code has been run the file get's deleted so that when new respondents would be included these would be added to the code.